Revert "Revert "Uniform component mass prior""
