+++ gen_summary.py
@@ -1,4 +1,4 @@
-from run_analysis import output_summary
+from parameter_estimation.save import saveSummary
 from os import listdir
 from os.path import isfile, join
 
@@ -8,7 +8,7 @@ from gwosc import datasets
 path = "output/posterior_samples"
 files = [f[:-3] for f in listdir(path) if isfile(join(path, f))]
 
-output_summary(files)
+saveSummary(files)
 
 gwtc3 = datasets.find_datasets(type='events', catalog='GWTC-3-confident')
 missing_events = set(gwtc3) - set(files)
