+++ parameter_estimation/plotting.py
@@ -1,18 +1,29 @@
+import corner
+import numpy as np
 import jax.numpy as jnp
 import jax
-
-import numpy as np
-import corner
-from pathlib import Path
 import matplotlib.pyplot as plt
 import h5py
-import os
 
-from pandas import DataFrame
-
+from .utilities import mkdir
 
 
-############################## Plot Posterior Samples ##############################
+"""
+Available params:
+['final_spin', 'spin_2y', 'final_mass_source', 'spin_1y', 'cos_tilt_2', 'mass_1_source', 
+'viewing_angle', 'spin_2x', 'inverted_mass_ratio', 'phi_2', 'chi_p', 'chirp_mass', 'chirp_mass_source', 
+'total_mass', 'redshift', 'luminosity_distance', 'theta_jn', 'chi_eff', 'a_1', 'cos_iota', 
+'radiated_energy', 'cos_theta_jn', 'total_mass_source', 'phi_jl', 'mass_2', 'ra', 'final_mass', 
+'spin_1x', 'log_likelihood', 'tilt_2', 'tilt_1', 'psi', 'dec', 'symmetric_mass_ratio', 'mass_2_source', 
+'iota', 'psi_J', 'cos_tilt_1', 'phi_12', 'mass_ratio', 'comoving_distance', 'phase', 'beta', 'chi_p_2spin', 
+'phi_1', 'a_2', 'spin_1z', 'peak_luminosity', 'spin_2z', 'mass_1', 'tilt_1_infinity_only_prec_avg', 
+'tilt_2_infinity_only_prec_avg', 'spin_1z_infinity_only_prec_avg', 'spin_2z_infinity_only_prec_avg', 
+'chi_eff_infinity_only_prec_avg', 'chi_p_infinity_only_prec_avg', 'cos_tilt_1_infinity_only_prec_avg', 
+'cos_tilt_2_infinity_only_prec_avg']
+"""
+
+############################## Functions for Plotting Various Graphs ##############################
+
 def plotPosterior(result, event, output_dir="output"):
     """
     Plot the posterior samples in a corner plot
@@ -25,23 +36,11 @@ def plotPosterior(result, event, output_dir="output"):
     figure = corner.corner(transposed_array, labels=labels, plot_datapoints=False, title_quantiles=[0.16, 0.5, 0.84], show_titles=True, title_fmt='g', use_math_text=True)
     mkdir(output_dir + "/posterior_plot")
     plt.savefig(output_dir + "/posterior_plot/"+event+".jpeg")
-
-
-############################## Save Posterior Samples ##############################
-def savePosterior(result, event, output_dir="output"):
+    
+    
+def plotRunningProgress(summary, event, output_dir="output"):
     """
-    Save the posterior sample points into a h5py file
-    """
-    samples = np.array(list(result.values())).reshape(15, -1) # flatten the array
-    transposed_array = samples.T # transpose the array
-    mkdir(output_dir + "/posterior_samples")
-    with h5py.File(output_dir + '/posterior_samples/' + event + '.h5', 'w') as f:
-        f.create_dataset('posterior', data=transposed_array)
-
-
-def plotRunAnalysis(summary, event, output_dir="output"):
-    """
-    Plot analysis of the run
+    Plot how the parameter estimation run progresses
     """
     chains, log_prob, local_accs, global_accs, loss_vals = summary.values()
     rng_key = jax.random.PRNGKey(42)
@@ -79,7 +78,8 @@ def plotRunAnalysis(summary, event, output_dir="output"):
     
     mkdir(output_dir + "/posterior_analysis")
     plt.savefig(output_dir + "/posterior_analysis/"+event+".jpeg")
-
+    
+    
 def plotLikelihood(summary, event, output_dir="output"):
     """
     Plot the likelihood of the run over epochs
@@ -101,32 +101,9 @@ def plotLikelihood(summary, event, output_dir="output"):
     plt.ylim(bottom=-20)
     mkdir(output_dir + "/likelihood")
     plt.savefig(output_dir + "/likelihood/"+event+".jpeg")
-
-
-def mkdir(path):
-    """
-    To create a directory if it does not exist
-    """
-    if not os.path.exists(path):
-        os.makedirs(path)
-
-
-"""
-Available params:
-['final_spin', 'spin_2y', 'final_mass_source', 'spin_1y', 'cos_tilt_2', 'mass_1_source', 
-'viewing_angle', 'spin_2x', 'inverted_mass_ratio', 'phi_2', 'chi_p', 'chirp_mass', 'chirp_mass_source', 
-'total_mass', 'redshift', 'luminosity_distance', 'theta_jn', 'chi_eff', 'a_1', 'cos_iota', 
-'radiated_energy', 'cos_theta_jn', 'total_mass_source', 'phi_jl', 'mass_2', 'ra', 'final_mass', 
-'spin_1x', 'log_likelihood', 'tilt_2', 'tilt_1', 'psi', 'dec', 'symmetric_mass_ratio', 'mass_2_source', 
-'iota', 'psi_J', 'cos_tilt_1', 'phi_12', 'mass_ratio', 'comoving_distance', 'phase', 'beta', 'chi_p_2spin', 
-'phi_1', 'a_2', 'spin_1z', 'peak_luminosity', 'spin_2z', 'mass_1', 'tilt_1_infinity_only_prec_avg', 
-'tilt_2_infinity_only_prec_avg', 'spin_1z_infinity_only_prec_avg', 'spin_2z_infinity_only_prec_avg', 
-'chi_eff_infinity_only_prec_avg', 'chi_p_infinity_only_prec_avg', 'cos_tilt_1_infinity_only_prec_avg', 
-'cos_tilt_2_infinity_only_prec_avg']
-"""
-
-
-def compare_plot(event_name, params, output_dir="compare_plot"):
+    
+    
+def plotCompare(event_name, params, output_dir="compare_plot"):
     """
     To compare the posterior samples from Jim and Bilby
     
@@ -215,7 +192,6 @@ def compare_plot(event_name, params, output_dir="compare_plot"):
     # spin_1_r = np.sqrt(spin_1x**2 + spin_1y**2 + spin_1z**2)
     # spin_2_r = np.sqrt(spin_2x**2 + spin_2y**2 + spin_2z**2)
 
-
     labels = params
 
     fig = corner.corner(jim_params, labels=labels, plot_datapoints=False, title_quantiles=[0.16, 0.5, 0.84], show_titles=True, title_fmt='g', use_math_text=True, color = 'red')
@@ -224,117 +200,15 @@ def compare_plot(event_name, params, output_dir="compare_plot"):
     plt.savefig(output_dir+"/"+event_name+".jpeg")
     
     
-
-def compare_intrinsic_params(event_name, output_dir="compare_plot"):
+def plotIntrinsicParamsComparison(event_name, output_dir="compare_plot"):
     """
     To compare the intrinsic parameters of the posterior samples from Jim and Bilby
     """
-    compare_plot(event_name, ["chirp_mass", "eta", "spin1x", "spin1y", "spin1z", "spin2x", "spin2y", "spin2z"], output_dir)
+    plotCompare(event_name, ["chirp_mass", "eta", "spin1x", "spin1y", "spin1z", "spin2x", "spin2y", "spin2z"], output_dir)
   
 
-def compare_extrinsic_params(event_name, output_dir="compare_plot"):
+def plotExtrinsicParamsComparison(event_name, output_dir="compare_plot"):
     """
     To compare the extrinsic parameters of the posterior samples from Jim and Bilby
     """
-    compare_plot(event_name, ["luminosity_distance", "phase", "iota", "psi", "ra", "dec"], output_dir)
-
-
-def plot_prior(event_name, params, output_dir="prior"):
-    """
-    params: A list of params to be included in the plot
-    """
-    bilby_posterior_dir = "data/IGWN-GWTC3p0-v1-" + event_name[:-3] + "_PEDataRelease_mixed_cosmo.h5"
-
-
-
-def KLdivergence(x, y):
-    """Compute the Kullback-Leibler divergence between two samples.
-    Parameters
-    ----------
-    x : 1D np array
-        Samples from distribution P, which typically represents the true
-        distribution.
-    y : 1D np rray
-        Samples from distribution Q, which typically represents the approximate
-        distribution.
-    Returns
-    -------
-    out : float
-        The estimated Kullback-Leibler divergence D(P||Q).
-    """ 
-    bins = np.linspace(min(np.min(x), np.min(y)), max(np.max(x), np.max(y)), 100) 
-    # The size of array must be much greater than the number of bins
-    
-    prob_x = np.histogram(x, bins=bins)[0]/x.shape
-    prob_y = np.histogram(y, bins=bins)[0]/y.shape
-       
-    return np.sum(np.where((prob_y != 0) & (prob_x != 0), prob_y * (np.log(prob_y) - np.log(prob_x)), 0))
-    
-    
-def test(x, y):
-    return (1+(0.5)**2)/(2)-0.5 
-
-
-def JSdivergence(P, Q):
-  #create M
-  M=(P+Q)/2 #sum the two distributions then get average
-
-  kl_p_q = KLdivergence(P, M)
-  kl_q_p = KLdivergence(Q, M)
-
-  js = (kl_p_q+kl_q_p)/2
-  return js
-
-
-def output_summary(events):
-    """
-    event: a list of events
-    """
-    
-    chirp_mass = []
-    standard_chirp_mass = []
-    chirp_mass_JS = []
-    
-    eta = []
-    standard_eta = []
-    eta_JS = []
-    
-    luminosity_distance = []
-    standard_luminosity_distance = []
-    luminosity_distance_JS = []
-    
-    for event in events:
-        print("Generating summary for "+event)
-        bilby_posterior_dir = "data/IGWN-GWTC3p0-v1-" + event[:-3] + "_PEDataRelease_mixed_cosmo.h5"
-        jim_posterior_dir = "output/posterior_samples/" + event + ".h5"    
-
-        file = h5py.File(jim_posterior_dir, 'r')
-        jim_posterior = np.array(file['posterior'])
-        file.close()
-
-        file = h5py.File(bilby_posterior_dir, 'r')
-        
-        jim_chirp_mass = jim_posterior[:,0]
-        chirp_mass.append(np.mean(jim_chirp_mass))
-        standard_chirp_mass.append(np.mean(np.array(file['C01:Mixed']['posterior_samples']['chirp_mass'])))
-        chirp_mass_JS.append(JSdivergence(np.random.choice(jim_chirp_mass, size=10000), np.random.choice(np.array(file['C01:Mixed']['posterior_samples']['chirp_mass']), size=10000)))
-        
-        jim_eta = jim_posterior[:,1]
-        eta.append(np.mean(jim_eta))
-        standard_eta.append(np.mean(np.array(file['C01:Mixed']['posterior_samples']['symmetric_mass_ratio'])))
-        eta_JS.append(JSdivergence(np.random.choice(jim_eta, size=10000), np.random.choice(np.array(file['C01:Mixed']['posterior_samples']['symmetric_mass_ratio']), size=10000)))
-        
-        jim_luminosity_distance = jim_posterior[:,8]
-        luminosity_distance.append(np.mean(jim_luminosity_distance))
-        standard_luminosity_distance.append(np.mean(np.array(file['C01:Mixed']['posterior_samples']['luminosity_distance'])))
-        luminosity_distance_JS.append(JSdivergence(np.random.choice(jim_luminosity_distance, size=10000), np.random.choice(np.array(file['C01:Mixed']['posterior_samples']['luminosity_distance']), size=10000)))
-        
-        file.close()
-    
-    df = DataFrame({'event': events, 'chirp_mass': chirp_mass, 'standard_chirp_mass': standard_chirp_mass, 'chirp_mass_JS': chirp_mass_JS, 'eta': eta, 'standard_eta': standard_eta, 'eta_JS': eta_JS, 'luminosity_distance': luminosity_distance, 'standard_luminosity_distance': standard_luminosity_distance, 'luminosity_distance_JS': luminosity_distance_JS})
-    df.to_excel('test.xlsx', sheet_name='pe_result', index=False)
-    
-
-
-
-
+    plotCompare(event_name, ["luminosity_distance", "phase", "iota", "psi", "ra", "dec"], output_dir)
