+++ run.sh
@@ -11,7 +11,7 @@ module load python cuda
 source /mnt/home/yxu10/venv_10/bin/activate
 cd $SLURM_SUBMIT_DIR
 
-python single_event_pe.py --event "GW191216_213338" --gps 1260567236.4 --duration 4 --Mc_prior 3.0 30.0 --ifos "H1" "V1" --heterodyned True
+python single_event_pe.py --event "GW191216_213338_uni" --gps 1260567236.4 --duration 4 --Mc_prior 3.0 30.0 --ifos "H1" "V1" --heterodyned True
 # python single_event_pe.py --event "GW200129_065458" --gps 1264316116.4 --duration 8 --Mc_prior 14.0 49.0 --ifos "H1" "L1" "V1"
 # python single_event_pe.py --event "GW200224_222234" --gps 1266618172.4 --duration 4 --Mc_prior 22.0 58.0 --ifos "H1" "L1" "V1"
 # python single_event_pe.py --event "GW200112_155838" --gps 1262879936.1 --duration 4 --Mc_prior 21.0 50.0 --ifos "L1" "V1"
