+++ single_event_pe.py
@@ -4,7 +4,7 @@ import jax
 import jax.numpy as jnp
 
 from jimgw.jim import Jim
-from jimgw.prior import Composite, Unconstrained_Uniform, Sphere
+from jimgw.prior import Composite, Unconstrained_Uniform, Sphere, UniformInComponentChirpMass, UniformInComponentMassRatio
 from jimgw.single_event.detector import H1, L1, V1
 from jimgw.single_event.likelihood import TransientLikelihoodFD, HeterodynedTransientLikelihoodFD
 from jimgw.single_event.waveform import RippleIMRPhenomPv2
@@ -15,7 +15,10 @@ from gwosc import datasets
 from tap import Tap
 import argparse
 
-from run_analysis import plotPosterior, savePosterior, plotRunAnalysis, plotLikelihood, mkdir
+from parameter_estimation.utilities import mkdir
+from parameter_estimation.plotting import plotPosterior, plotRunningProgress, plotLikelihood
+from parameter_estimation.save import savePosterior
+
 
 
 
@@ -54,14 +57,20 @@ def runSingleEventPE(output_dir, event, gps, duration, post_trigger_duration, Mc
     ###########################################
     ########## Set up priors ##################
     ###########################################
-
-    Mc_prior = Unconstrained_Uniform(Mc_prior[0], Mc_prior[1], naming=["M_c"])
-    q_prior = Unconstrained_Uniform(
-        0.125,
-        1.0,
+    # Mc_prior = Unconstrained_Uniform(Mc_prior[0], Mc_prior[1], naming=["M_c"])
+    # q_prior = Unconstrained_Uniform(
+    #     0.125,
+    #     1.0,
+    #     naming=["q"],
+    #     transforms={"q": ("eta", lambda params: params["q"] / (1 + params["q"]) ** 2)},
+    # )
+    Mc_prior = UniformInComponentChirpMass(Mc_prior[0], Mc_prior[1], naming=["M_c"])
+    q_prior = UniformInComponentMassRatio(
+        0.125, 
+        1.0, 
         naming=["q"],
         transforms={"q": ("eta", lambda params: params["q"] / (1 + params["q"]) ** 2)},
-    )
+        )
     s1_prior = Sphere(naming="s1")
     s2_prior = Sphere(naming="s2")
     dL_prior = Unconstrained_Uniform(0.0, 10000.0, naming=["d_L"])
@@ -189,7 +198,7 @@ def runSingleEventPE(output_dir, event, gps, duration, post_trigger_duration, Mc
     mkdir(output_dir)
     plotPosterior(result, event, output_dir)
     savePosterior(result, event, output_dir)
-    plotRunAnalysis(summary, event, output_dir)
+    plotRunningProgress(summary, event, output_dir)
     plotLikelihood(summary, event, output_dir)
 
 
@@ -218,7 +227,6 @@ if __name__ == "__main__":
     parser.add_argument('--waveform', type=str, default="RippleIMRPhenomPv2", help='Your name')
     parser.add_argument('--heterodyned', type=bool, default=False, help='Your name')
     args = parser.parse_args()
-    
     
     runSingleEventPE(
         output_dir=args.output_dir,
